LS: Move project backend modules to a new form
This is a separate commit to make diffs of the next one smaller,
as Git was not detecting file renames.

commit-id:02b4cddc
