+++ crates/cairo-lang-language-server/src/lib.rs
@@ -94,8 +94,8 @@ use crate::config::Config;
 use crate::ide::semantic_highlighting::SemanticTokenKind;
 use crate::lang::diagnostics::lsp::map_cairo_diagnostics_to_lsp;
 use crate::lang::lsp::LsProtoGroup;
-use crate::project::backends::cairo_project::corelib::try_to_init_unmanaged_core;
-use crate::project::backends::scarb::db::update_crate_roots;
+use crate::project::scarb::db::update_crate_roots;
+use crate::project::unmanaged_core_crate::try_to_init_unmanaged_core;
 use crate::project::ProjectManifestPath;
 use crate::server::notifier::Notifier;
 use crate::toolchain::scarb::ScarbToolchain;
