--- crates/cairo-lang-language-server/src/project/backends/cairo_project/mod.rs
@@ -1,2 +1,0 @@
-// TODO(mkaput): This is `pub` temporarily.
-pub(crate) mod corelib;
