--- crates/cairo-lang-language-server/src/project/backends/mod.rs
@@ -1,4 +1,0 @@
-// TODO(mkaput): This is `pub` temporarily.
-pub(crate) mod cairo_project;
-// TODO(mkaput): This is `pub` temporarily.
-pub(crate) mod scarb;
